backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
google-generativeai>=0.3.2
pillow>=10.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
pypdf2>=3.0.1
reportlab>=4.2.0
openpyxl>=3.1.2
bcrypt>=4.0.0
pymupdf>=1.24.0